build: include components.d.ts
This file has been gitignored earlier because the keys in the autogenerated types would keep changing their order and creating unnecessary modifications. However, excluding it from the tracked files made it necessary to rebuild the entire project in every postinstall, which could take a while, and that's exactly what we'd like to avoid when running CI/CD pipelines. Other StencilJS projects include this file, and I figured we should do that as well while looking for a proper solution as opposed to simply ignoring the whole thing.