test(foxy-profile): fix timeout trigger
Replacing waitForEvent with waitFor(ms) because sometimes the target event is fired before the listener is registered, causing the test to fail with a timeout error. The timeout of 1 second should be enough because no real requests are made in this case – every api call is intercepted and responded to immediately.